+++ CHANGELOG.md
@@ -7,6 +7,11 @@ GAMSPy 0.11.0
 - General
   - Generate expression representation as soon as it is created to avoid tall recursions.
   - Find variables in equations by iteratively traversing instead of doing recursion.
+  - Add NEOS Server as a backend to solve models.
+- Documentation
+  - Add documentation for NEOS backend.
+- Testing
+  - Add tests for NEOS backend.
 
 -------------------------------------------------------------------------------
 GAMSPy 0.10.5